fixed author bug

+++ app/controllers/news_controller.rb
@@ -1,12 +1,12 @@
 class NewsController < ApplicationController
   def home
-  	@articles = Article.find(:all, :order => "id desc", :limit => 8)
+    @articles = Article.find(:all, :order => "id desc", :limit => 8)
 
 
   end
 
   def article
-  	@article = Article.find(params[:id])
+    @article = Article.find(params[:id])
 
 
 
@@ -18,7 +18,7 @@ class NewsController < ApplicationController
 
 
   end
-  def delete 
+  def delete
     if params[:id] and @user and @user.author
       @article = Article.find(params[:id])
       @article.destroy
@@ -28,37 +28,38 @@ class NewsController < ApplicationController
     end
   end
   def insert
-  	@article = Article.new
+    if @user
+      @article = Article.new
 
-  	@article.title = params[:title]
-  	@article.body = params[:body]
-  	@article.person = Person.find(1)
+      @article.title = params[:title]
+      @article.body = params[:body]
+      @article.person_id = @user.id
 
-  	if @article.save and @article.title != "" and @article.body != ""
-      Reference.where(:article_id => @article.id).each do |d|
-        d.destroy
-      end
-      if params[:persons]
-        params[:persons].each do |p|
-          r = Reference.new
-          r.person_id = p
-          r.article_id = @article.id
-          r.save
+      if @article.save and @article.title != "" and @article.body != ""
+        Reference.where(:article_id => @article.id).each do |d|
+          d.destroy
         end
-      end
-      if params[:groups]
-        params[:groups].each do |g|
-          r = Reference.new
-          r.group_id = g
-          r.article_id = @article.id
-          r.save
+        if params[:persons]
+          params[:persons].each do |p|
+            r = Reference.new
+            r.person_id = p
+            r.article_id = @article.id
+            r.save
+          end
         end
+        if params[:groups]
+          params[:groups].each do |g|
+            r = Reference.new
+            r.group_id = g
+            r.article_id = @article.id
+            r.save
+          end
+        end
+        redirect_to "/article/"+ @article.id.to_s
+      else
+        redirect_to "/error"
       end
-  		redirect_to "/article/"+ @article.id.to_s
-  	else
-  		redirect_to "/error"
-  	end
-
+    end
   end
 
 end
